requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
beautifulsoup4==4.12.3
blinker==1.9.0
blobfile==3.0.0
cachetools==5.5.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.8
clipboard==0.0.4
DateTime==5.5
distro==1.9.0
dotenv==0.9.9
filelock==3.16.1
fire==0.7.0
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.1
idna==3.10
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
llama_models==0.0.63
llama_stack==0.0.63
llama_stack_client==0.0.63
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.1.0
multitasking==0.0.11
narwhals==1.27.1
nltk==3.9.1
numpy==2.1.3
ollama==0.4.7
packaging==24.2
pandas==2.2.3
peewee==3.17.8
pillow==11.1.0
platformdirs==4.3.6
plotly==6.0.0
prompt_toolkit==3.0.48
propcache==0.3.0
protobuf==5.29.3
pyaml==25.1.0
pyarrow==19.0.1
pycryptodomex==3.21.0
pydantic==2.10.5
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyvenv==0.2.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.23.1
setuptools==75.8.0
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
streamlit==1.42.2
tenacity==9.0.0
termcolor==1.1.0
textblob==0.19.0
tiktoken==0.8.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
webencodings==0.5.1
yarl==1.18.3
yfinance==0.2.50
zope.interface==7.2
